Move `LazySmpPool` into its own module. Bench: 5797963
