+++ src/protocols/mod.rs
@@ -1,5 +1,6 @@
 //! External interfaces for UCI and XBoard communication.
 
+pub mod smp;
 pub mod spmc;
 pub mod uci;
 pub mod xboard;
